Feito a funcionalidade de adicionar a hora e recuperar do banco na MainActivity, feito a validação de hora igual nos campos, alinhado a hora e minuto na MainActivity, "Toda a hora adicionada ao banco é convertida em minutos e ai sim gravado na base de dados, na hora que se recupera, o sistema transforma em horas e minutos"

+++ app/src/main/java/com/br/horasestudos/views/business/DisciplineBusiness.java
@@ -34,7 +34,7 @@ public class DisciplineBusiness {
         return repository.loadDiscipline(id);
     }
 
-    public Boolean insertHour(int id, String data, float total) {
+    public Boolean insertHour(int id, String data, long total) {
         try {
             SQLiteDatabase db = MainDb.getInstance().getWritableDatabase();
             String query = ("UPDATE Discipline SET date = '" + data + "', all_hours = all_hours + '" + total + "' WHERE ID = '" + id + "'");
